actualizacion enableJpaRepositories 7yy

+++ hotels/src/main/java/com/co/buritica/hotels/repositories/CitiesRepository.java
@@ -3,5 +3,5 @@ package com.co.buritica.hotels.repositories;
 import com.co.buritica.hotels.models.entities.CitiesEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CitiesRepository extends JpaRepository <CitiesEntity, Integer>{
+public interface CitiesRepository extends JpaRepository <CitiesEntity, Long>{
 }
